feat(feature): close darkbar improvement and rearranging folders
